Halve the memory a loaded track costs (TD-09 part)
A deck holds a whole decoded track, which makes the sample format the app's
largest memory cost by a wide margin. Stored as i16 rather than f32, a
five-minute track at 48 kHz drops from 110 MB to 55, and the two decks loaded
near a transition from 220 MB to 110.

Nothing is lost that was in the recording. The sources are lossy AAC and MP3
whose own noise floor sits far above 16 bit, and the conversion divides by 32768
rather than 32767 so it is an exact power of two and round-trips losslessly. It
clamps before scaling, so a sample above full scale saturates instead of wrapping
to the opposite sign and turning a hot master into a burst of noise — a test
pins both properties.

The stretcher reads i16 and converts on the way out, once per emitted frame. Its
correlation search now runs on integers, which is also cheaper than it was. Beat
alignment is unchanged: the rendered-onset test still passes at the same
tolerance, which is the measurement that would have caught a timing regression.

Streaming is still the real fix and is not done. It would take a deck to about
1 MB, but it needs a decoder thread, a windowed source the WSOLA search can
index backwards into, and seek semantics that wait for the producer. Attempting
that at the end of a long session is how dropouts get reintroduced into a path
that has been careful about exactly that, so it is left as the next focused
piece of work with the constraints written down.

Web tier scope decided: full parity. The browser build analyses and plays for
itself, with nothing deferred to a desktop instance — a web tier that phoned
home to a desktop would contradict the claim that the device in front of you
does the work, and there is no sync mechanism to build that split on anyway.

+++ vapor-core/crates/vapor-engine/src/deck.rs
@@ -3,6 +3,14 @@
 //! Mirrors one `AudioStreamPlayer` + its `DeckA`/`DeckB` bus in the Godot
 //! build: source audio, a stretch ratio, an EQ/filter chain, and a gain.
 //!
+//! ## Samples are stored as `i16`
+//!
+//! A deck holds a whole decoded track, which makes the sample format the app's
+//! largest memory cost: a five-minute track at 48 kHz is 110 MB as stereo `f32`
+//! and 55 MB as `i16`, and two decks are loaded near a transition. The sources
+//! are lossy AAC and MP3 whose own noise floor sits far above 16 bit, so this
+//! discards nothing that was in the recording.
+//!
 //! The critical structural difference: Godot polled `get_playback_position()`
 //! every frame from three separate `_process` loops and drove bus parameters
 //! through tweens on the main thread. Here the deck advances by exactly the
@@ -23,7 +31,7 @@ const FREEZE_ROOM: f32 = 0.95;
 const FREEZE_DAMPING: f32 = 0.1;
 
 pub struct Deck {
-    samples: Vec<[f32; 2]>,
+    samples: Vec<[i16; 2]>,
     sample_rate: f32,
     stretcher: Stretcher,
     eq: EqChain,
@@ -70,7 +78,7 @@ impl Deck {
         }
     }
 
-    pub fn load(&mut self, samples: Vec<[f32; 2]>) {
+    pub fn load(&mut self, samples: Vec<[i16; 2]>) {
         drop(self.swap_samples(samples));
     }
 
@@ -85,7 +93,7 @@ impl Deck {
     /// Returning it lets the caller move the buffer back to a control thread
     /// and drop it there. Everything else about the two is identical, so
     /// `load` is written in terms of this rather than beside it.
-    pub fn swap_samples(&mut self, samples: Vec<[f32; 2]>) -> Vec<[f32; 2]> {
+    pub fn swap_samples(&mut self, samples: Vec<[i16; 2]>) -> Vec<[i16; 2]> {
         let previous = std::mem::replace(&mut self.samples, samples);
         self.stretcher.reset(0.0);
         self.eq.reset();
